Work with error when close

+++ core/entities/Ini_parser/entities/Ini_parser.cpp
@@ -1,6 +1,5 @@
 #include "Ini_parser.hpp"
 
-
 Ini_parser::Ini_parser(const std::string &filename) {
     const std::vector<std::string> lines = libio::file::read_file(filename);
     this->m_sections = parse_ini_from_vector(lines);
@@ -14,7 +13,7 @@ Ini_parser::Sections_t Ini_parser::get_sections() const {
     return this->m_sections;
 }
 
-int Ini_parser::get_section_count() const {
+unsigned long Ini_parser::get_section_count() const {
     return this->m_sections.size();
 }
